+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - `Innmind\Http\Header\AcceptRanges` constructor is now private, use `::of()` or `::maybe()` named constructors
 - `Innmind\Http\Header\Age` constructor is now private, use `::of()` or `::maybe()` named constructors
 - `Innmind\Http\Header\Authorization` constructor is now private, use `::of()` or `::maybe()` named constructors
+- `Innmind\Http\Header\ContentEncoding` constructor is now private, use `::of()` or `::maybe()` named constructors
 
 ### Removed
 
@@ -34,6 +35,7 @@
 - `Innmind\Http\Header\AcceptRangesValue`
 - `Innmind\Http\Header\AgeValue`
 - `Innmind\Http\Header\AuthorizationValue`
+- `Innmind\Http\Header\ContentEncodingValue`
 
 ### Fixed